Implementation of the data dumping on the filesystem. Part 1.
The files are currently stored in the directory from which the application
is run. The next step will consist in selecting a folder.

+++ core_prj/core/BaseCamera.h
@@ -17,7 +17,7 @@ class CORESHARED_EXPORT BaseCamera : public QObject
     Q_OBJECT
 
 public:
-    static const int size = 2048;
+    static const qint64 size = 2048;
 
     typedef quint16 Pixel;
     typedef Pixel Snapshot[size][size] ;
